requirements.txt
# Flask and Related Packages
Flask>=2.3.0
Flask-Cors>=4.0.0
Werkzeug>=3.0.0
flask-cors 
flask-limiter

# Llama Index and Related Packages
llama-index==0.11.13
llama-index-cli==0.3.1
llama-index-core==0.11.13.post1
llama-index-llms-ollama==0.3.2
llama-index-embeddings-ollama==0.3.1
llama-index-llms-groq

# Data Science and Visualization
numpy==1.26.4
pandas==2.2.3
matplotlib==3.9.2

# Environment and Database
python-dotenv
SQLAlchemy

# Miscellaneous
pyvis